Add protection against potential edge-case bug
If the dict users is somehow empty, the else clause would result in attempting to call .delete() on None.

+++ cogs/autochess_cog.py
@@ -389,7 +389,8 @@ class AutoChessCog(BaseCog):
             # of getting banned for scraping OP.GG's website
             await asyncio.sleep(10)
         else:
-            await msg.delete()
+            if msg:
+                await msg.delete()
         
         if arg in ALL_ARGS:
             u = "all users"
